+++ src/include/AST/Native.h
@@ -33,6 +33,9 @@
 namespace AST {
 
 /// The 'native` type.
+/// 
+/// \invariant kind () == Item::Kind::NATIVE
+/// 
 class Native :
 	public ItemWithId
 {